+++ src/main/java/Javatro/Manager/Options/HelpIntroOption.java
@@ -14,17 +14,17 @@ public class HelpIntroOption implements Option {
     /** Executes the command to display the introduction message. */
     @Override
     public void execute() {
-        System.out.println("\n=== Welcome to Javatro ===");
-        System.out.println(
-                "Javatro is a text-based roguelike deck-building game inspired by Balatro.");
-        System.out.println(
-                "Your goal is to strategically build a deck and progress through challenges.");
-        System.out.println(
-                "Each turn, you'll play poker-style hands to defeat enemies and earn new cards.");
-        System.out.println(
-                "\nThink strategically, manage your deck wisely, and see how far you can go!");
+//        System.out.println("\n=== Welcome to Javatro ===");
+//        System.out.println(
+//                "Javatro is a text-based roguelike deck-building game inspired by Balatro.");
+//        System.out.println(
+//                "Your goal is to strategically build a deck and progress through challenges.");
+//        System.out.println(
+//                "Each turn, you'll play poker-style hands to defeat enemies and earn new cards.");
+//        System.out.println(
+//                "\nThink strategically, manage your deck wisely, and see how far you can go!");
 
-        JavatroManager.setScreen(UI.getHelpScreen());
+        JavatroManager.setScreen(UI.getHelpIntroScreen());
     }
 
     /**
